AVRO-744. C: Helper macros for extracting and setting record field values
This patch adds two helper macros that can be used to extract or set the
value of a record field using fewer function calls.  It wraps a call to
avro_record_get, which extracts the avro_datum_t instance for the field,
and a call to avro_XXX_get or avro_XXX_set on the field datum.


git-svn-id: https://svn.apache.org/repos/asf/avro/trunk@1070941 13f79535-47bb-0310-9956-ffa450edef68