AVRO-396. C: avrocat and avropipe commands
The new avrocat command prints the contents of an Avro file to stdout.
Each record appears on a separate line, and is formatted using the Avro
JSON encoding.

The new avropipe command prints the same contents, but the output format
is the same as if you had run the JSON encoding of the data through the
jsonpipe [1] command.  This provides a format that's more amenable to
processing using standard Unix tools, since individual scalar values (no
matter how deep the original record) appear on separate lines.

[1] https://github.com/dvxhouse/jsonpipe


git-svn-id: https://svn.apache.org/repos/asf/avro/trunk@1146936 13f79535-47bb-0310-9956-ffa450edef68