AVRO-1034. C: Resolved reader initializes child array.
The resolved reader array implementation now initializes its children
array.  We only need to initialize the elements of the array when
they're first created.  This happens the first time we access a
particular element, or any element that appears after it in the array.

Test case contributed by Vivek Nadkarni.


git-svn-id: https://svn.apache.org/repos/asf/avro/trunk@1295404 13f79535-47bb-0310-9956-ffa450edef68