AVRO-1047: Java: Fix remaining javac warnings.
AVRO-1913 fixed most warnings, but missed a few unchecked usages and
left in some redundant casts.  This patch eliminates remaining
warnings even with the -Xlint:all flag added to javac.  With no more
warnings to surpress, the @SuppressWarnings("all") annotation is
removed from the generated code, as well.  The TestSpecificCompiler
test has been modified to pass -Xlint:all to the compiler and assert
that no warnings are present in compiled code.