Writer module overhauling (#20)
* Make encode and decode private

There is little reason why these should be public, especially if we
expose a write function with very little overhead.

They can always be made public when needed.

* Substitute SingleWriter with to_avro_datum

The avro implementations in other languages provide a DataFileWriter and
a DatumWriter which are respectively the interface to write to a file
with buffering and headers and the interface to just convert a value to
avro.

While Writer is definitely equivalent (actually more generic than
DataFileWriter since it accepts anything implementing Write), the
SingleWriter struct wasn't really similar to DatumWriter. For this
reason we replaced SingleWriter with the function to_avro_datum which
does exactly what DatumWriter does.

Please notice that the Rust package doesn't provide a public
Encoder/Decoder interface like other implementations at the moment,
since we only support binary encoding. Given that we don't even pass the
writable interface down to the encoder (because we can't see why we
should do it), this makes the Encoder/Decoder pretty much useless for
the users of the library.

This commit also changes the visibility rules of the writer module,
re-exporting Writer and to_avro_datum at the lib level.

* Documentation for writer module

* Add unit tests to writer module

The number of bytes returned by the writing operations didn't add up
with the length of the output data. Fixed two bugs causing that.

* Implement Writer::extend_ser for parity with extend

Until we are not able to make Value implement Serialize, we are obliged
to duplicate the interface for the writer.