[AVRO-2226] Fixes UnionSchema specificity
Trouble arises in the python library when deducing the appropriate schema from a list of schemas given a particular datum.

When "null" values are allowed for fields in two separate schemas, there is no way to differentiate which should schema be used given of a list of schemas that are set as the type definition for a record.

This PR checks to ensure all fields defined on a given datum are _also_ defined in the schema being validated to use for that datum.

With this bugfix, datums such as `{"foo": "a"}` will not "cast" to schemas such as `{"name": "bar", "type": ["long", "null"]}`, which is currently the case.