AVRO-2112: Target .NET Standard/Core in C# (#307)
* AVRO-2161: Upgrade C# unit tests to NUnit 3

* AVRO-2112: Update C# projects to target .NET Standard/Core and .NET Framework 4.0

* AVRO-2112: Move C# IPC tests to a Avro.ipc.test

* AVRO-2112: Ignore C# tests that use System.CodeDom compilation when targeting .NET Core

See https://github.com/dotnet/corefx/issues/12180

* AVRO-2112: Replace usage of JToken.ToString() in C# projects

In Newtonsoft.Json v3.5, JToken.ToString() returned the raw JSON
representation of the token. In later versions of
Newtonsoft.Json, JToken.ToString() returns a simple string
representation of the value. See the examples below:
- v3.5: "\"Hello World\""
- Later versions: "Hello World"

In this commit, I've updated the project to work with later versions
of Newtonsoft.Json as well as v3.5. I've replaced some usages of
JToken.ToString(). When we need the raw JSON representation, we use
JsonConvert.Serialize(). When we need the string value of a string
JToken, we use JToken.Value<string>().

* AVRO-2112: Update C# README

* Update pom.xml with new C# paths to ignore for license check

* csharp: Cut support for net35, update Dockerfile and csharp build.sh

* Add mono-complete to Dockerfile for build

* csharp: Remove unnecessary ProjectGuid elements from projects

* Revert to only running tests for Avro.test.dll

This is what we were doing on master. I was trying to run the IPC tests,
but that has not seemed to work well in our automated builds.

* List all tests as they are run

* Add build.ps1 and update README

* Update Target Frameworks table in README

* Try disabling tests that take a long time to run