AVRO-2515: Expand access of ProtobufData methods (#612)
Chandged from private to public getSchema(FieldDescriptor) and
getNamespace(FileDescriptor, Descriptor).

The utility methods that deal with the conversion from Protobuf message
descriptor to an Avro schema are now exposed on a more granular level.
This will alow the library clients to use this functionality and not
reinvent it. This will also help with extending the base functionality.
For example, adding protobuf meta information to the schema as
properties.