AVRO-831 Refactor lang/py setup and test structure (#733)
* AVRO-831: Pythonic Build

* AVRO-831: Rework lang/py Setup

Decouple python setup from ant and enable tests to run without it.

* AVRO-831: Run Tests Normally

* AVRO-831: Skip Java Tether if No JDK

Make it easier to test on systems without Java

* AVRO-831: Modern Exception Syntax

* AVRO-831: Exterminate Ants

* AVRO-831: Found One More Ant

* AVRO-831: Remove pdb

* AVRO-831: iSort Order

* AVRO-831: Rename Build Subcommand

* AVRO-831: Correct PYTHONPATH in Interop Tests

* AVRO-831: Skip Test if JAR Not Found

* AVRO-831: Set Package Version Correctly

Fixes an error where the package metadata version is not used, causing setuptools to ignore it.

Co-Authored-By: RyanSkraba <ryan@skraba.com>

* AVRO-831: Ignore More Files Created by Setup

Adds additional files that are created during setup to .gitignore because they are generated and should not be in version control.

* AVRO-831: Clean Directly from build.sh

Python packaging is [moving away from embedding commands in
setup.py][1]. It is difficult to maintain external commands in Python
this way. Managing the dependencies needed to run commands from within
setup.py is gnarly, because dependencies cannot be resolved so early.
Furthermore, it's difficult to test code that happens at the point at
which tests themselves are triggered, so that code needs to be very
simple. A shell script is simple and more appropriate for this use case.

At another time we should look to extract the lint command as well.

[1][https://github.com/pypa/setuptools/issues/931]