AVRO-2652: Use Tox to Enable Multi Python Testing (#742)
* AVRO-2652: Fix Inconsistent Pretty Output Order

Using json.dumps on a dict gives the simplest possible way to
pretty-print a schema, but it doesn't give any control over the output
order. But the test kit assumes the output is in a specific order, which
is not guaranteed by older Python implementations. This change sorts the
keys for the simplest possible stable output order.

* AVRO-2652: Run Tests with Tox

Tox enables testing with multiple versions of python in the same place.

* AVRO-2652: Tox in Docker

* AVRO-2657: Fix Python Interop Data Generator

* AVRO-2657: Test with Interop Data

* AVRO-2657: Make build.sh do interop tests