AVRO-1938: Generate Parsing Canonical Forms of Schema (#1167)
* AVRO-1938: Generate Parsing Canonical Forms of Schema

This PR adds support for generating Parsing Canonical Forms of Avro Schemas
to the main `avro` package.

The bulk of work was done by @kojiromike and @forsberg. This PR cleans up code
where necessary, adds more test cases, and clarifies on transformations where
not applicable in Python (ex. Transformation of integers with leading zeros)

Closes: https://issues.apache.org/jira/browse/AVRO-1938

* AVRO-1938 Fix typo of repeating test method name

* AVRO-1938: Makes `names` argument to `to_canonical_json` optional

This commit ensures that `names` is an optional parameter in all subclass
implementations of `to_canonical_json`. It also addresses minor formatting
issues and review comments to previous commits.

* AVRO-1938 Use `dict.get` to return value in dict or default