AVRO-3252: Add hasattr check for mode attr in DataFileReader (#1406)
The DataFileWriter class constructor checks for the presence of a mode attribute before accessing it, but the DataFileReader does not. This means that it is possible to use the DataFileWriter to write data to in memory binary streams using io.BytesIO, but it is not possible to read from them in the same manner. This commit adds the same check to the DataFileReader class.