[C++] Run clang-format on all files (#2928)
There is a .clang-format file in the repository, indicating that files
should be formatted with this standard, but they have drifted out of
conformance over time, which makes it easy to accidentally introduce
unrelated differences when submitting changes. This formats the entire
c++ directory.

Ideally, format compliance should be checked when a pull request is
made, I'm open to creating a PR which adds this workflow if desired.